Add a test case for #27
